API/wwwroot/scripts/requirements.txt
click==8.1.7
opencv-python==4.10.0.82
pillow==10.3.0
PyPDF2==3.0.1
PyMuPDF==1.24.3
reportlab==4.2.0
numpy==1.26.4
